+++ Changelog.md
@@ -29,6 +29,9 @@ Ideas and To-Dos:
     * To process
 * Export python scripts to d/l satellite imagery to match
 ___________________________________________________________
+2021-05-24: updated to master
+* Add uncertainties to SeaBASS Es & Rrs output
+
 ---
 v1.0.8 2021-05-11: DAA